[Minor] Code simplification, clean up and more type details
