+++ spec/Provider/RemainingTotalProviderSpec.php
@@ -17,16 +17,15 @@ use PhpSpec\ObjectBehavior;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
 use Sylius\RefundPlugin\Entity\RefundInterface;
 use Sylius\RefundPlugin\Model\RefundType;
-use Sylius\RefundPlugin\Provider\DelegatingRefundUnitTotalProviderInterface;
+use Sylius\RefundPlugin\Provider\RefundUnitTotalProviderInterface;
 use Sylius\RefundPlugin\Provider\RemainingTotalProviderInterface;
+use Symfony\Contracts\Service\ServiceProviderInterface;
 
 final class RemainingTotalProviderSpec extends ObjectBehavior
 {
-    function let(
-        DelegatingRefundUnitTotalProviderInterface $refundUnitTotalProvider,
-        RepositoryInterface $refundRepository,
-    ): void {
-        $this->beConstructedWith($refundUnitTotalProvider, $refundRepository);
+    function let(ServiceProviderInterface $refundUnitTotalProviders, RepositoryInterface $refundRepository): void
+    {
+        $this->beConstructedWith($refundUnitTotalProviders, $refundRepository);
     }
 
     function it_implements_remaining_total_provider_interface(): void
@@ -35,13 +34,16 @@ final class RemainingTotalProviderSpec extends ObjectBehavior
     }
 
     function it_returns_remaining_total_to_refund(
-        DelegatingRefundUnitTotalProviderInterface $refundUnitTotalProvider,
+        ServiceProviderInterface $refundUnitTotalProviders,
+        RefundUnitTotalProviderInterface $refundUnitTotalProvider,
         RepositoryInterface $refundRepository,
         RefundInterface $refund,
     ): void {
         $refundType = RefundType::orderItemUnit();
 
-        $refundUnitTotalProvider->getRefundUnitTotal(1, $refundType)->willReturn(1000);
+        $refundUnitTotalProviders->get($refundType->getValue())->willReturn($refundUnitTotalProvider);
+        $refundUnitTotalProvider->getRefundUnitTotal(1)->willReturn(1000);
+
         $refundRepository
             ->findBy(['refundedUnitId' => 1, 'type' => $refundType->__toString()])
             ->willReturn([$refund])
@@ -53,12 +55,14 @@ final class RemainingTotalProviderSpec extends ObjectBehavior
     }
 
     function it_returns_unit_total_if_there_is_no_refund_for_this_unit_yet(
-        DelegatingRefundUnitTotalProviderInterface $refundUnitTotalProvider,
+        ServiceProviderInterface $refundUnitTotalProviders,
+        RefundUnitTotalProviderInterface $refundUnitTotalProvider,
         RepositoryInterface $refundRepository,
     ): void {
         $refundType = RefundType::orderItemUnit();
 
-        $refundUnitTotalProvider->getRefundUnitTotal(1, $refundType)->willReturn(1000);
+        $refundUnitTotalProviders->get($refundType->getValue())->willReturn($refundUnitTotalProvider);
+        $refundUnitTotalProvider->getRefundUnitTotal(1)->willReturn(1000);
 
         $refundRepository
             ->findBy(['refundedUnitId' => 1, 'type' => $refundType->__toString()])
