--- src/Provider/DelegatingRefundUnitTotalProviderInterface.php
@@ -1,21 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Paweł Jędrzejewski
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Sylius\RefundPlugin\Provider;
-
-use Sylius\RefundPlugin\Model\RefundTypeInterface;
-
-interface DelegatingRefundUnitTotalProviderInterface
-{
-    public function getRefundUnitTotal(int $id, RefundTypeInterface $refundType): int;
-}
